test(events): hoist EventServiceBundle into a shared fixture
Five files under tests/events/ each rebuild the same EventServiceBundle
inline. Hoist the construction into a new tests/events/conftest.py
under the fixture name ``event_services`` (matching the top-level
``services`` fixture).

- test_handlers_asset.py, test_handlers_maintenance_record.py: drop
  the local ``event_services`` fixture.
- test_pagination.py: drop the local ``bundle`` fixture and rename
  the test-parameter to ``event_services``.
- test_catchup_cross_tenant_isolation.py: drop the ``_bundle()``
  factory and take ``event_services`` as a fixture parameter.
- test_event_log_type_id.py: take ``event_services`` instead of
  constructing inline.

tests/events/test_bundle.py keeps its local fixture because it
substitutes a counting spy for AssetTypeFieldService.
