+++ tests/events/test_pagination.py
@@ -7,7 +7,6 @@ from uuid import uuid4
 
 import pytest
 
-from novamoc.domain.events._bundle import EventServiceBundle
 from novamoc.domain.events._pagination import EventLogCursorPaginator
 from novamoc.domain.events._payloads import (
     Created,
@@ -16,10 +15,6 @@ from novamoc.domain.events._payloads import (
     RecordedEvent,
 )
 from novamoc.domain.events.services import EventLogService
-from novamoc.domain.schema.services import (
-    AssetTypeFieldService,
-    MaintenanceRecordTypeFieldService,
-)
 from tests.data.scenarios import ACTIVE_TRUCK
 
 if TYPE_CHECKING:
@@ -28,6 +23,7 @@ if TYPE_CHECKING:
 
     from sqlalchemy.ext.asyncio import AsyncSession
 
+    from novamoc.domain.events._bundle import EventServiceBundle
     from tests.data.scenarios import Scenario
 
 
@@ -50,18 +46,6 @@ def paginator(session: AsyncSession) -> EventLogCursorPaginator:
     return EventLogCursorPaginator(EventLogService(session=session))
 
 
-@pytest.fixture
-def bundle(session: AsyncSession) -> EventServiceBundle:
-    return EventServiceBundle(
-        asset_type_field_service=AssetTypeFieldService(session=session),
-        maintenance_record_type_field_service=MaintenanceRecordTypeFieldService(
-            session=session
-        ),
-        event_log_service=EventLogService(session=session),
-        schema_version=0,
-    )
-
-
 async def test_get_items_empty_stream_returns_no_items_and_no_cursor(
     paginator: EventLogCursorPaginator,
 ) -> None:
@@ -72,11 +56,11 @@ async def test_get_items_empty_stream_returns_no_items_and_no_cursor(
 
 async def test_get_items_returns_all_when_under_page_size(
     paginator: EventLogCursorPaginator,
-    bundle: EventServiceBundle,
+    event_services: EventServiceBundle,
     seed: Callable[[Scenario], Awaitable[Mapping[str, Mapping[str, UUID]]]],
 ) -> None:
     type_id = (await seed(ACTIVE_TRUCK))["asset_type"]["Truck"]
-    await _append_n(bundle, 3, type_id)
+    await _append_n(event_services, 3, type_id)
     items, cursor = await paginator.get_items(cursor=None, results_per_page=10)
     assert len(items) == 3
     assert all(isinstance(it, RecordedEvent) for it in items)
@@ -86,11 +70,11 @@ async def test_get_items_returns_all_when_under_page_size(
 
 async def test_get_items_returns_first_page_and_signals_more(
     paginator: EventLogCursorPaginator,
-    bundle: EventServiceBundle,
+    event_services: EventServiceBundle,
     seed: Callable[[Scenario], Awaitable[Mapping[str, Mapping[str, UUID]]]],
 ) -> None:
     type_id = (await seed(ACTIVE_TRUCK))["asset_type"]["Truck"]
-    await _append_n(bundle, 5, type_id)
+    await _append_n(event_services, 5, type_id)
     items, cursor = await paginator.get_items(cursor=None, results_per_page=2)
     assert len(items) == 2
     assert cursor == items[-1].seq
@@ -98,11 +82,11 @@ async def test_get_items_returns_first_page_and_signals_more(
 
 async def test_get_items_cursor_handoff_continues_stream(
     paginator: EventLogCursorPaginator,
-    bundle: EventServiceBundle,
+    event_services: EventServiceBundle,
     seed: Callable[[Scenario], Awaitable[Mapping[str, Mapping[str, UUID]]]],
 ) -> None:
     type_id = (await seed(ACTIVE_TRUCK))["asset_type"]["Truck"]
-    await _append_n(bundle, 5, type_id)
+    await _append_n(event_services, 5, type_id)
     page1, cursor1 = await paginator.get_items(cursor=None, results_per_page=2)
     page2, cursor2 = await paginator.get_items(cursor=cursor1, results_per_page=2)
     page3, cursor3 = await paginator.get_items(cursor=cursor2, results_per_page=2)
@@ -119,11 +103,11 @@ async def test_get_items_cursor_handoff_continues_stream(
 
 async def test_get_items_exact_page_boundary_signals_caught_up(
     paginator: EventLogCursorPaginator,
-    bundle: EventServiceBundle,
+    event_services: EventServiceBundle,
     seed: Callable[[Scenario], Awaitable[Mapping[str, Mapping[str, UUID]]]],
 ) -> None:
     type_id = (await seed(ACTIVE_TRUCK))["asset_type"]["Truck"]
-    await _append_n(bundle, 4, type_id)
+    await _append_n(event_services, 4, type_id)
     items, cursor = await paginator.get_items(cursor=None, results_per_page=4)
     assert len(items) == 4
     assert cursor is None  # the +1 fetch returned only 4, so we're done
